Token Creation Time reloading
When vault resets the Token after a certain TTL (by default : 10 minutes), The CreationTime Variable must also be reloaded. Otherwise Vault will reset again and again Token after the first 10 minutes. This was tested in a large scale environement and it turned out to be a problem because Hashicorp does not like to create too many tokens.